+++ StockBoxData/changelog.md
@@ -2,6 +2,13 @@
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/).
+## [1.0.1] - 2024-05-09
+
+### Changed
+- BaseIndicator now uses generic type for `Payload` property and return type of `CalculateIndicator` method to allow children to define their payloads
+
+### Added
+- PriceChannel indicator
 
 ## [1.0.0] - 2024-05-08