+++ StockBox_UnitTests/changelog.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [1.0.1] - 2023-05-09
+### Changed
+- cast all indicators to access typed payload
+
+### Added
+- unit test for new PriceChannel indicator
+
 ## [1.0.0] - 2024-05-08
 
 ### Added